Add Firestore service and update dependencies
Added new Firestore service files and updated the necessary dependencies in `pubspec.yaml` and related platform-specific files. This includes changes to `GeneratedPluginRegistrant.swift` for macOS and `generated_plugin_registrant.cc` and `generated_plugins.cmake` for Windows. Also, updated the `pubspec.lock` file to reflect the new versions of the dependencies.